rename triangle to triangles, to avoid conflict names with triangle library
