+++ gdsfactory/components/__init__.py
@@ -220,7 +220,7 @@ from gdsfactory.components.text_rectangular import (
     text_rectangular,
     text_rectangular_multi_layer,
 )
-from gdsfactory.components.triangle import triangle, triangle2, triangle4
+from gdsfactory.components.triangles import triangle, triangle2, triangle4
 from gdsfactory.components.verniers import verniers
 from gdsfactory.components.version_stamp import pixel, qrcode, version_stamp
 from gdsfactory.components.via import via, via1, via2, viac
