+++ gdsfactory/components/dicing_lane.py
@@ -1,7 +1,7 @@
 import gdsfactory as gf
 from gdsfactory.component import Component
 from gdsfactory.components.rectangle import rectangle
-from gdsfactory.components.triangle import triangle
+from gdsfactory.components.triangles import triangle
 from gdsfactory.types import ComponentSpec, Float2, LayerSpec
 
 triangle_metal = gf.partial(triangle, layer="M3", xtop=2)
